Minor improvements

+++ pde/fields/vectorial.py
@@ -114,7 +114,9 @@ class VectorField(DataFieldBase):
         # evaluate all vector components at all points
         data = []
         for expression in expressions:
-            expr = ScalarExpression(expression=expression, signature=grid.axes)
+            expr = ScalarExpression(
+                expression=expression, signature=grid.axes, dtype=dtype
+            )
             values = np.broadcast_to(expr(**points), grid.shape)
             data.append(values)
